Simulate generating pubkey_package from dealer
