+++ dealer/Cargo.toml
@@ -4,9 +4,20 @@ version = "1.0.0"
 edition = "2021"
 
 [dependencies]
+gateway = { path = "../gateway" }
 frost-ed25519 = "2.1.0"
 rand = "0.8" 
 serde = { workspace = true }
-tokio = { version = "1", default-features = false, features = ["macros", "sync", "time", "rt-multi-thread", "signal"] }
-anyhow = "1"
+tokio = { workspace = true }
+libp2p = { version = "0.53", features=["rendezvous", "async-std", "dns", "gossipsub", "mdns", "noise", "macros", "request-response", "tcp", "tokio", "yamux","identify", "ping", "cbor", "serde"] }
+libp2p-mplex = "0.41.0"
+libp2p-quic = { version = "0.10", features=["tokio"] }
 bincode = { workspace = true }
+anyhow = "1"
+clap = { workspace = true }
+serde_json = "1"
+async-trait = "0.1"
+bytes = "1"
+
+
+
